Fix network ports datatype (unsigned int -> unsigned short). UDPSocket: fix old bug using m_port instead of m_port[x].

+++ P25Gateway/P25Network.cpp
@@ -24,7 +24,7 @@
 #include <cassert>
 #include <cstring>
 
-CP25Network::CP25Network(unsigned int port, const std::string& callsign, bool debug) :
+CP25Network::CP25Network(unsigned short port, const std::string& callsign, bool debug) :
 m_callsign(callsign),
 m_socket(),
 m_port(port),
