+++ P25Reflector/Conf.h
@@ -45,7 +45,7 @@ public:
   bool         getLogFileRotate() const;
 
   // The Network section
-  unsigned int getNetworkPort() const;
+  unsigned short getNetworkPort() const;
   bool         getNetworkDebug() const;
 
 private:
@@ -61,7 +61,7 @@ private:
   std::string  m_logFileRoot;
   bool         m_logFileRotate;
 
-  unsigned int m_networkPort;
+  unsigned short m_networkPort;
   bool         m_networkDebug;
 };
